Fix folder structure for About page
